MB-10533 Move `LoginTaskSet` from `tasks/base` to `utils/task`
This way more of our custom helpers live in utils rather than in other directories.
